app/requirements/base.txt
autopep8==1.5.5
boto3==1.17.25
celery==4.4.6
dj-database-url==0.5.0
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
flower==0.9.4
packaging==20.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
redis==3.5.3
sendgrid==6.6.0
whitenoise==5.2.0
email-validator==1.1.2
channels==3.0.3
channels-redis==3.2.0
pytest-asyncio==0.14.0
pytest-django==4.1.0
django-redis==4.12.1
locust==1.2.2
pylint==2.6.0
pyfcm==1.4.7
pyotp==2.4.1
sentry-sdk==1.0.0
django-import-export==2.5.0
twilio==6.53.0
Faker==6.6.0
selenium==3.141.0
django-graphql-jwt==0.3.1 
graphene==2.1.8 
graphene-django==2.15.0
django-graphene-permissions==0.0.4
graphene-file-upload==1.2.2
graphene-django-extras==0.4.9
django-graphql-social-auth==0.1.4
django-channels-graphql-ws==0.8.0